DESAFIOS/desafio_bots/requirements.txt
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bleach==3.1.0
blinker==1.4
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
ConfigArgParse==0.15.1
crypto==1.4.1
cryptography==2.8
docker==4.1.0
docker-compose==1.25.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.1
Flask-JWT-Extended==3.24.1
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
locust==0.0
locustio==0.13.5
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.1.0
msgpack-python==0.5.6
Naked==0.1.31
paramiko==2.7.1
pkginfo==1.5.0.1
pyasn1==0.4.8
pycparser==2.19
pycrypto==2.6.1
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyrsistent==0.15.7
python-anticaptcha==0.4.2
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
sentry-sdk==0.12.3
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.5
texttable==1.6.2
urllib3==1.25.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
zipp==2.0.0
